Clean up file structure for multiple codemods and bootstrap injectGlobalToCreateGlobalStyle
