+++ src/v4/injectGlobalToCreateGlobalStyle.js
@@ -1,0 +1,8 @@
+"use strict";
+
+const t = require("babel-core").types;
+const _ = require("lodash");
+
+module.exports = function() {
+  return {};
+};
